Turned on more warnings in the compiler and fixed their causes or suppressed them.

+++ dont_distribute/mic/cor_feat_sel/src/edu/wright/cs/birg/CSVUtil.java
@@ -138,8 +138,8 @@ public final class CSVUtil {
 		
 		if(header == null){
 			if(len != null){
-				header = new String[len];
-				for(int i = 0; i < len; ++i){
+				header = new String[len.intValue()];
+				for(int i = 0; i < len.intValue(); ++i){
 					header[i]="Col"+i;
 				}
 			}else{
